+++ lib/action.ts
@@ -1,0 +1,44 @@
+"use server"
+
+import { auth } from "@clerk/nextjs/server";
+import { z } from "zod";
+import { prisma } from "./prisma";
+
+export const addPostAction = async (formData: FormData) => {
+
+    try {
+        const { userId } = auth();
+        if (!userId) {
+            console.log('userId is null')
+            return;
+        }
+
+        const postText = formData.get("post") as string
+        const postTextSchema = z.string().min(1, '入力OK？').max(140, '140字以内')
+        const validatedPostText = postTextSchema.parse(postText)
+
+        await prisma.post.create({
+            data: {
+                content: validatedPostText,
+                authorId: userId,
+            }
+        })
+    } catch (error) {
+        if (error instanceof z.ZodError) {
+            return {
+                error: error.errors.map((e) => e.message).join(", "),
+                success: false
+            }
+        } else if (error instanceof Error) {
+            return {
+                error: error.message,
+                success: false
+            }
+        } else {
+            return {
+                error: '予期せぬエラー',
+                success: false
+            }
+        }
+    }
+}
